+++ CHANGELOG.md
@@ -59,6 +59,8 @@ Kaye Chat *Dify App*:
   when only provided `difficulty_override`,
   (sense node is still required for sensing PLs)
 
+- include Annotation Markers for coder
+